requirements.txt
Django==4.0
django-cors-headers==3.10.1
djangorestframework==3.13.0